+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Features
+- Add chevron icon next to project title to highlight that it is actionable.
 - Create initial (sandbox) project and activity when creating a new user.
 - Include project and activity usage in project list item.